Create text input for transaction amount
The display for users to input the transaction amount has been
created. It currently uses a similar UI to that of PayNow as the
user can input only numbers and it will ignore any non-numerical
keypresses. It also forces the user to start typing from the second
decimal place. All this is to ensure that the user does not input
erroneous values, e.g. numbers containing multiple decimal points,
commas, etc.

The enum for TransactionType has been shifted into Transaction.ts
for easier reference.