Complete create transaction feature
Creation of transactions has been completed and now will update the
list of transactions upon creation of new transactions. UI will
also update accordingly. Transactions are sorted according to the
date of transaction from most recent to least recent for easier
reference.

The UI for creating transactions has also been shifted from a tab
to a Model to allow useState modifiers to be passed into it to
update the UI upon creating a new transaction.