Fix line graph bug
Line Graph component was causing some errors as the data arrays
were initialized as empty arrays. It is now changed to initialize
as an array of length 1.

Calendar now has 6 rows instead of 5 to accommodate for months that
span 6 weeks.

A navigation bar is also implemented to enable the user to navigate
across months.

Headers are also added to the calendar to indicate the day of the
week.