Create touchable calendar
Calendar component now supports the user to click on any of the
dates to view all transactions on the date.

Line graph component has been shifted to another page which will
contain more statistics using other types of representation, e.g.
pie charts. To be decided on a future date.

It is known that if the sum of transactions is too big of a number,
the text will wrap within the grid, so something should be done to
tackle this in the future.