Resolve Calendar UI issue
Previously the height for individual CalendarGrids are hard coded
to be a percentage of the screen's height and width. It has been
modified to use the built in 'height' and 'width' components of the
StyleSheet to allow to it to resize automatically according to the
size of its container.

Also starting on using a pie chart to show the breakdown of the
month's transactions by type.