Change Transaction class implementation
Previously Transaction class used an enum to store the different types
of transactions, but that did not allow for users to add new types of
transactions as they deem fit. Thus a new class TransactionType has been
created to store all the different types of transactions in a string
array, which then allows the user to add or remove types.

However, a known limitation of this implementation is that the user can
create a transaction with a type, e.g. 'food', and subsequently delete
that type of transaction. Thus it will lead to a transaction with a non
existent type. Despite this, I think it is acceptable for now as it
shouldn't affect the functionality of the app.

Also the pie chart has been adjusted to no longer be cut off due to the
offset that was used previously.