Enable navigation across months
Previously the user could not move to the next month when viewing
transactions in the monthly summary page. Now user can move to the next
month when they are on the last day of the month.

i.e. if user is on 31 Jan 2024, clicking next on the modal will bring
them to 1 Feb 2024, displaying all transactions on that day.