Improve Create Transaction UX
Previously the user had to manually select the date each time they
wanted to create a transaction record, which is pretty cumbersome when
they want to create multiple records. Now the user can click on a
specific date and the new transaction record's default date will be the
selected date.